ENH: minor HLC fix

+++ turtlebot_proj_nav/src/HighLevelCommand.cpp
@@ -193,7 +193,7 @@ void HighLevelCommand::sendDistanceAndAngleCommand(const float linearVelocity, c
 void HighLevelCommand::seekMarker()
 {
     std::cout<<commandBusy<<std::endl;
-    std::cout<<(!commandBusy.data) && (disableCommand.data == false)<<std::endl;
+    std::cout<<(int)((!(commandBusy.data)) && (disableCommand.data == false))<<std::endl;
     if ((!commandBusy.data) && (disableCommand.data == false))
          {
          switch (seekingMarkerState)
